+++ engine/src/test/java/net/jqwik/api/ExhaustiveGenerationTests.java
@@ -246,18 +246,6 @@ class ExhaustiveGenerationTests {
 		}
 
 		@Example
-		@Deprecated
-		@Label("Arbitraries.samples() returns all samples in row")
-		void samples() {
-			Optional<ExhaustiveGenerator<String>> optionalGenerator = Arbitraries.samples("a", "b", "c").exhaustive();
-			assertThat(optionalGenerator).isPresent();
-
-			ExhaustiveGenerator<String> generator = optionalGenerator.get();
-			assertThat(generator.maxCount()).isEqualTo(3);
-			assertThat(generator).containsExactly("a", "b", "c");
-		}
-
-		@Example
 		@Label("Arbitraries.frequency() returns all in row")
 		void frequency() {
 			Tuple.Tuple2<Integer, String> frequency1 = Tuple.of(1, "a");